refactor: project modularity
