+++ .vscode/settings.json
@@ -60,6 +60,13 @@
         "array": "cpp",
         "deque": "cpp",
         "span": "cpp",
-        "text_encoding": "cpp"
+        "text_encoding": "cpp",
+        "cwctype": "cpp",
+        "functional": "cpp",
+        "memory_resource": "cpp",
+        "numeric": "cpp",
+        "random": "cpp",
+        "numbers": "cpp",
+        "variant": "cpp"
     }
 }
