+++ include/auth/delete_session.h
@@ -1,0 +1,6 @@
+#ifndef DELETE_SESSION_H
+#define DELETE_SESSION_H
+
+void deleteSession();
+
+#endif
