+++ include/auth/get_latest_userid.h
@@ -1,0 +1,6 @@
+#ifndef GET_LATEST_USERID_H
+#define GET_LATEST_USERID_H
+
+int getLatestUserId();
+
+#endif
