+++ include/auth/get_session.h
@@ -1,0 +1,9 @@
+#ifndef GET_SESSION_H
+#define GET_SESSION_H
+
+#include <string>
+#include <vector>
+using namespace std;
+vector<string> getSession();
+
+#endif
