+++ include/auth/login_menu.h
@@ -1,0 +1,6 @@
+#ifndef LOGIN_MENU_H
+#define LOGIN_MENU_H
+
+int loginMenu();
+
+#endif
