+++ include/auth/login_signup_menu.h
@@ -1,0 +1,6 @@
+#ifndef LOGIN_SIGNUP_MENU_H
+#define LOGIN_SIGNUP_MENU_H
+
+void loginSignupMenu();
+
+#endif
