+++ include/auth/register_menu.h
@@ -1,0 +1,6 @@
+#ifndef REGISTER_MENU_H
+#define REGISTER_MENU_H
+
+int registerMenu();
+
+#endif
