+++ include/recipe/add_recipe.h
@@ -1,0 +1,6 @@
+#ifndef ADD_RECIPE_H
+#define ADD_RECIPE_H
+
+void addRecipe();
+
+#endif
