+++ include/recipe/delete_recipe.h
@@ -1,0 +1,6 @@
+#ifndef DELETE_RECIPE_H
+#define DELETE_RECIPE_H
+
+void deleteRecipe();
+
+#endif
