+++ include/recipe/edit_recipe.h
@@ -1,0 +1,6 @@
+#ifndef EDIT_RECIPE_H
+#define EDIT_RECIPE_H
+
+void editRecipe();
+
+#endif
