+++ include/recipe/find_recipe.h
@@ -1,0 +1,6 @@
+#ifndef FIND_RECIPE_H
+#define FIND_RECIPE_H
+
+void findRecipe();
+
+#endif
