+++ include/recipe/recipe_not_found.h
@@ -1,0 +1,6 @@
+#ifndef RECIPE_NOT_FOUND_H
+#define RECIPE_NOT_FOUND_H
+
+void recipeNotFound();
+
+#endif
