+++ include/utils/clear_screen.h
@@ -1,0 +1,6 @@
+#ifndef CLEAR_SCREEN_H
+#define CLEAR_SCREEN_H
+
+void clearScreen();
+
+#endif
