+++ src/auth/get_latest_userid.cpp
@@ -1,0 +1,33 @@
+#include "get_latest_userid.h"
+#include <iostream>
+#include <fstream>
+#include <sstream>
+using namespace std;
+
+int getLatestUserId()
+{
+    ifstream file("./database/users.csv");
+    string line;
+    int maxId = 0;
+
+    if (!file) {
+        cout << "Basis data tidak ditemukan!" << endl;
+        return 0;
+    }
+
+    while (getline(file, line))
+    {
+        stringstream ss(line);
+        string idStr;
+        getline(ss, idStr, ',');
+        
+        int id = stoi(idStr); // stoi: string to integer
+        if (id > maxId) {
+            maxId = id;
+        }
+    }
+
+    int nextId = maxId + 1;
+
+    return nextId;
+}
