+++ src/auth/login_signup_menu.cpp
@@ -1,0 +1,22 @@
+#include "login_signup_menu.h"
+#include <iostream>
+using namespace std;
+
+void loginSignupMenu()
+{
+    cout << "\n";
+    cout << "===========================" << endl;
+    cout << "|                         |" << endl;
+    cout << "| Selamat datang di       |" << endl;
+    cout << "| Food Note!              |" << endl;
+    cout << "| Silahkan masuk/daftar   |" << endl;
+    cout << "|                         |" << endl;
+    cout << "===========================" << endl;
+    cout << "|                         |" << endl;
+    cout << "|  [1] Masuk              |" << endl;
+    cout << "|  [2] Daftar             |" << endl;
+    cout << "|  [3] Hentikan Program   |" << endl;
+    cout << "|                         |" << endl;
+    cout << "===========================" << endl;
+    cout << "Navigasi ke halaman > ";
+}
