+++ src/recipe/show_main_menu.cpp
@@ -1,0 +1,23 @@
+#include "show_main_menu.h"
+#include <iostream>
+using namespace std;
+
+void showMainMenu()
+{
+    cout << "\n";
+    cout << "========================" << endl;
+    cout << "|                      |" << endl;
+    cout << "|      Food Note       |" << endl;
+    cout << "|                      |" << endl;
+    cout << "========================" << endl;
+    cout << "|                      |" << endl;
+    cout << "| [1] Buat Resep Baru  |" << endl;
+    cout << "| [2] Daftar Resep     |" << endl;
+    cout << "| [3] Cari Resep       |" << endl;
+    cout << "| [4] Edit Resep       |" << endl;
+    cout << "| [5] Hapus Resep      |" << endl;
+    cout << "| [0] Keluar           |" << endl;
+    cout << "|                      |" << endl;
+    cout << "------------------------" << endl;
+    cout << "Navigasi ke halaman > ";
+}
