+++ src/utils/clear_screen.cpp
@@ -1,0 +1,13 @@
+#include "clear_screen.h"
+#include <iostream>
+using namespace std;
+
+// menghapus screen
+void clearScreen ()
+{
+    #ifdef _WIN32
+        system("cls"); // Windows
+    #else
+        system("clear"); // Linux / Mac
+    #endif
+}
